2.x

+++ blossom-module-crm/blossom-module-crm-biz/src/main/java/cn/econets/blossom/module/crm/dal/dataobject/product/CrmProductDO.java
@@ -1,6 +1,7 @@
 package cn.econets.blossom.module.crm.dal.dataobject.product;
 
 import cn.econets.blossom.framework.mybatis.core.dataobject.BaseDO;
+import cn.econets.blossom.module.crm.enums.DictTypeConstants;
 import cn.econets.blossom.module.crm.enums.product.CrmProductStatusEnum;
 import com.baomidou.mybatisplus.annotation.KeySequence;
 import com.baomidou.mybatisplus.annotation.TableId;
@@ -37,13 +38,13 @@ public class CrmProductDO extends BaseDO {
     /**
      * 单位
      *
-     * 字典 {@link cn.econets.blossom.module.crm.enums.DictTypeConstants#CRM_PRODUCT_UNIT}
+     * 字典 {@link DictTypeConstants#CRM_PRODUCT_UNIT}
      */
     private Integer unit;
     /**
      * 价格，单位：分
      */
-    private Long price;
+    private Integer price;
     /**
      * 状态
      *
